Change the CURRENT_CONFIGURATION just before calling to Fortran's codegen
The main problem was that we were changing the CURRENT_CONFIGURATION before
adding the new file to the compilation process. Note that when we add a new file,
we propagate some flags from the current configuration to the one that will be used to compile that file.
Thus, we were not properly propagating those flags.


It fixes #2803