Fix ETDG constants computation for nested loops
When loops are nested and the inner loop boundaries depend on the outer loop's induction variable, then there are several reaching definitions arriving to the point where the inner induction variable must be computed. So, reaching definitions cannot be used for such a purpose.
This commit uses the information computed for loops boundaries top compute constants in such cases.