Do not assume that codegen is able to fix this case
When walking through a FunctionCall in the Codegen's phase, if the
implicit function form is present, Codegen will ignore the function call
and traverses the first argument of the call.

The problematic situation is when we are emitting the initializer of
a symbol definition: in that case, when the initializer is an implicit function call,
we traverse all the arguments, not just the 1st one. Basically because in some
situations we generate implicit function calls where all the arguments must be emmited.