Make sure we emit the VLA symbols found inside a sizeof expression
It is possible to specify a type-id in a sizeof expression with an array
declarator with nonconstant size.

A similar issue had happened in the past with cast expression which was
solved implementing a stack of extra declarations which are flushed at the
statement level.

A similar change should be done with __alignof__ but we're hitting #2863.
So let's fix first the common sizeof case and then we can fix the much
less common case of __alignof__ later.