This test is only valid in OmpSs-2
The test introduced in this merge-req is not supported by our Nanos++ lowering, thus it should not be executed in OmpSs mode. I would move it to the OmpSs-2 directory.

The reason why is not supported was explained in the ticket that added support to iterators that depend on the value of other iterators (new feature introduced in #2850). Basically, the main problem is that in Nanos++ we need to know the total number of dependences to properly allocate a Nanos++ internal data structure. Thus, we would need to traverse twice the iteration space defined by our iterators to know the dynamic number of dependences we have for a certain task.

See #2868