Remove --typecheck
This flag was useful when we first implemented typechecking as it was
easy to slip error types without diagnostics. I don't remember being
useful recentlu.

This loses us a bit of checking during testing. If we believe this is
unacceptable I suggest we create an extra check after the FE (similar to
ast_check) that runs only when no diagnostic has been emitted. The pass
should check that no tree has an error type.