Make a failure to end with a signal
Unless the test is marked with test_{compile,exec}_end_signal (with the
usual allowed variants in the same style as fail, faulty), ending with a
signal is always an error.

This way we can tell a test that fails and ends normally apart from a
test that ends with a signal (like abort or segmentation fault).