Allow decltype in a base-specifier
This is an oversight when we added support of C++11, decltypes can be
used as base-specifiers in classes. Similar to nested-name specifiers,
this is of limited use as the expression in the decltype has to
ultimately be a class type.

If the expression is type-dependent, we will wrap it in a SK_DECLTYPE.
We need to update codegen and the instantiation of bases so it takes
this case into account.