Fix size calculation
We are replacing the set of three expressions (1, i, i + 1) using a
range [lower, upper). This means that the extended set of expressions
must look like (upper - lower, lower, upper).

However this expression is difficult to express just as a direct
substitution of the original.  So we defined "upper1 ← upper - 1" and
used it as (upper1 - lower, lower, upper1 + 1).

The third expression is effectively "upper1 + 1" == upper - 1 + 1 == upper".

The first expression, though, "upper1 - lower" is not equivalent to
"upper - lower". We need to compensate it as "upper1 - lower + 1"
so it is "upper - 1 - lower + 1".