Do not update the size
The size of the object doesn't change just because we are extending its
dependency, so don't touch it.

There are some corner cases, like a dependency of the form `p[i:i]` (for a
pointer `T *p`) where it would be arguable to update it. But for now we
don't seem to have a realistic use case for those (also Nanos6 doesn't
use the last dimension for anything now).