Do not regenerate a file that is an input of gperf
tpp generates a file that we later use as an input of gperf,
accidentally triggering a regeneration of one gperf-generated file.

In the past (though this does not happen anymore because we changed the
host that creates the distributable of Mercurium) this means that a
single gperf-generated file was built with the host gperf. This gperf
may have different characteristics of the gperf used to create the
distributable. For long there were not any differences but as of gperf
3.1 files may differ with earlier versions in an incompatible way.

Because we may not want to trigger a rebuild of gperf-generated files,
we need to avoid tpp generating that input of gperf.

Missing tools like gperf, flex and bison prevents developing Mercurium
itself, so let's add a warning when this happens: the build is useable
for a compiler user but not a compiler developer.