Fix gperf dependences
Make sure we regenerate all gperf outputs if we build
gperf-compat-types.h. Otherwise it may happen that mismatching gperf
versions (e.g 3.0 vs 3.1) cause build failures.

The typical build failure happens in this case
 - we use a host with gperf 3.0 to "make dist"
 - we use a host with gperf 3.1 to build from a tar.gz (but without
   running make clean first)

This causes gperf-compat-types.h to be updated for gperf-3.1 but
distributed files are still for gperf-3.0. This causes a compilation
error due to a conflicting type.

To fix this, make sure we run gperf when we update
gperf-compat-types.h, so add a dependency in the Makefile.