Fix lambda helper generation
Currently we lower the C++11 lambdas explicitly in Mercurium, but this
means we need to be very careful when the new definitions are emitted.

Due to the recursive nature of the lambdas, they were being emitted too
early and this ended in the wrong place. By chance codegen emitted the
code correctly even if the nodecl was wrong.

This does not fix all the issues (see #2970) but fix the cases that seem
the most common ones.