OpenACC: Changed the parameter appending strategy to also include task function declarations
We found a bug in cases where an openacc device task is declared in a header file
but the source file using it doesn't have the actual definition. This caused the
declaration from the header to not be appropriately modified for the OpenACC async
argument, while the calls were modified correctly, resulting in compilation errors.