Enable PGI compiler detection out of PATH.
It is sometimes needed to be able to use PGI-enabled Mercurium profiles
without having PGI in the path, or its module loaded in the case of BSC clusters.

This commit introduces a --with-pgi-installation config flag, to manually specify
a PGI installation. The compilers detected in it will be used with their full path.