Allow an expression as a label argument
LLVM/OmpSs-2 will only support 'const char*' in its label clause (fixing
a long-running mistake of Mercurium).

While we can't really change what Mercurium uses internally as a label,
we can allow using string literals rather than just a bunch of random
tokens. This would make this particular form of the label clause
compatible both in Mercurium and LLVM/OmpSs-2.

The old syntax is supported but a warning is emitted suggesting
the better syntax.

test.c:10:9: warning: deprecated use of label(foo bar)
             will be parsed as label("foo bar")

The only corner cases that won't be supported now are

  label("foo)

but this often emits a warning in the C preprocessor, so this should
be regarded as something questionable that we don't have to support.